Resolved crash when not connected on startup
Now displays a "Not Connected" string when started with no connectivity.

Possible future feature would be a manual refresh which clears the
cache and pulls down fresh images.
Curgently when connectivity restored it doesn't automatically pull down
the maps, but if the user changes the pin or map size then the images
are pulled down, except for the back map isn't pulled until the user
clears the cache.